test(entity-currency-fiat): split registry tests into constants/internals/registry
